--- content_service/src/main/java/api/user/service/UserContentService.java
@@ -1,4 +1,0 @@
-package api.user.service;
-
-public class UserContentService {
-}
