--- content_service/src/main/java/common/exception/user/UserContentException.java
@@ -1,9 +1,0 @@
-package common.exception.user;
-
-import common.exception.ContentException;
-
-public class UserContentException extends ContentException {
-    public UserContentException(String message) {
-        super(message);
-    }
-}
