+++ content_service/src/main/java/urdego/common/exception/user/UserContentException.java
@@ -1,0 +1,10 @@
+package urdego.common.exception.user;
+
+import urdego.common.exception.ContentException;
+import urdego.common.exception.ExceptionMessage;
+
+public class UserContentException extends ContentException {
+    public UserContentException(ExceptionMessage message) {
+        super(message.getText());
+    }
+}
