+++ content_service/src/main/java/urdego/domain/entity/BaseEntity.java
@@ -1,4 +1,4 @@
-package domain.entity;
+package urdego.domain.entity;
 
 
 import jakarta.persistence.Column;
