+++ content_service/src/main/java/urdego/domain/entity/user/repository/UserContentRepository.java
@@ -1,6 +1,6 @@
-package domain.entity.user.repository;
+package urdego.domain.entity.user.repository;
 
-import domain.entity.user.UserContent;
+import urdego.domain.entity.user.UserContent;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserContentRepository extends JpaRepository<UserContent, Long> {
